+++ libs/react/src/__mocks__/categoriesMock.ts
@@ -474,3 +474,211 @@ export function getFootprintDataFacilitiesAllFootprintsNull() {
   }
   return data;
 }
+
+export function getFootprintAllNullFootprintValues() {
+  return {
+    "idx": 0,
+    "category_name": "Company Decarbonization",
+    "subcategories": {
+        "travel": {
+            "idx": 1,
+            "activities": {
+                "commuting": {
+                    "activity_name": "Commuting",
+                    "activity_description": "Employee transportation to/from worksite (ICE, EV, bus, train, scooter, bike, walking, etc.) "
+                },
+                "air_travel": {
+                    "activity_name": "Air Travel",
+                    "activity_description": "Business-related air travel, domestic and international. "
+                },
+                "other_travel": {
+                    "activity_name": "Other Travel",
+                    "activity_description": "Rail, boat, taxi, and car service travel, hotel and boarding costs, food and expenses, etc."
+                },
+                "vehicle_fuel": {
+                    "footprint": {
+                        "2022": {
+                            "value": null,
+                            "period_type": "year"
+                        },
+                        "2023": {
+                            "value": null,
+                            "period_type": "year"
+                        },
+                        "2023-q4": {
+                            "value": null,
+                            "period_type": "quarter"
+                        }
+                    },
+                    "activity_name": "Vehicle Fuel",
+                    "activity_description": "Sales, Executive, Delivery, and Operational Fleet Vehicles either owned or leased"
+                },
+                "vehicle_electricity": {
+                    "activity_name": "Vehicle Electricity",
+                    "activity_description": "Sales, Executive, Delivery, and Operational Fleet Vehicles either owned or leased"
+                }
+            },
+            "journey_score": 22,
+            "subcategory_name": "Travel"
+        },
+        "product": {
+            "idx": 3,
+            "activities": {
+                "materials": {
+                    "activity_name": "Materials",
+                    "activity_description": "Purchased raw materials and manufactured inputs."
+                },
+                "product_use": {
+                    "activity_name": "Product Use",
+                    "activity_description": "Optimize product use impact through customer education and behavior change.  "
+                },
+                "packaging_disposal": {
+                    "activity_name": "Packaging Disposal",
+                    "activity_description": "Consumer disposal of packaging (recycling, composting, trash, etc.) "
+                },
+                "manufacturing_waste": {
+                    "activity_name": "Manufacturing Waste",
+                    "activity_description": "Reduce production-related materials waste. "
+                },
+                "packaging_materials": {
+                    "activity_name": "Packaging Materials",
+                    "activity_description": "What raw materials + manufactured products are purchased for packaging. "
+                },
+                "manufacturing_energy": {
+                    "footprint": {
+                        "2022": {
+                            "value": null,
+                            "period_type": "year"
+                        },
+                        "2023": {
+                            "value": null,
+                            "period_type": "year"
+                        },
+                        "2023-q4": {
+                            "value": null,
+                            "period_type": "quarter"
+                        }
+                    },
+                    "activity_name": "Manufacturing Energy",
+                    "activity_description": "Process energy (electric + heat) used for making products."
+                },
+                "engineering_and_innovation": {
+                    "activity_name": "Engineering And Innovation",
+                    "activity_description": "Design new products and services for environmental impact."
+                },
+                "upstream_shipping_and_freight": {
+                    "activity_name": "Upstream Shipping And Freight",
+                    "activity_description": "Movement of materials and manufactured inputs to manufacturing facilities (includes import freight by land, air, or sea).  "
+                },
+                "downstream_shipping_and_freight": {
+                    "activity_name": "Downstream Shipping And Freight",
+                    "activity_description": "Movement of finished product to distributors, retailers, and customers by land, sea, or air."
+                }
+            },
+            "journey_score": null,
+            "subcategory_name": "Product"
+        },
+        "facilities": {
+            "idx": 0,
+            "activities": {
+                "fuel_owned": {
+                    "activity_name": "Fuel Owned",
+                    "activity_description": "Natural Gas, Propane, and Diesel consumption at all company owned facilities"
+                },
+                "fuel_leased": {
+                    "activity_name": "Fuel Leased",
+                    "activity_description": "Natural Gas, Propane, and Diesel consumption at all company leased facilities"
+                },
+                "electricity_owned": {
+                    "footprint": {
+                        "2022": {
+                            "value": null,
+                            "period_type": "year"
+                        },
+                        "2023": {
+                            "value": null,
+                            "period_type": "year"
+                        },
+                        "2023-q4": {
+                            "value": null,
+                            "period_type": "quarter"
+                        }
+                    },
+                    "activity_name": "Electricity Owned",
+                    "activity_description": "Electricity consumption at all company owned facilities"
+                },
+                "electricity_leased": {
+                    "footprint": {
+                        "2022": {
+                            "value": null,
+                            "period_type": "year"
+                        },
+                        "2023": {
+                            "value": null,
+                            "period_type": "year"
+                        },
+                        "2023-q4": {
+                            "value": null,
+                            "period_type": "quarter"
+                        }
+                    },
+                    "activity_name": "Electricity Leased",
+                    "activity_description": "Electricity consumption at all company leased facilities"
+                }
+            },
+            "journey_score": 15,
+            "subcategory_name": "Facilities"
+        },
+        "operations": {
+            "idx": 2,
+            "activities": {
+                "furniture": {
+                    "activity_name": "Furniture",
+                    "activity_description": "Furniture for offices, company-owned or leased residences, and other worksites.  "
+                },
+                "machinery_fuel": {
+                    "footprint": {
+                        "2022": {
+                            "value": null,
+                            "period_type": "year"
+                        },
+                        "2023": {
+                            "value": null,
+                            "period_type": "year"
+                        },
+                        "2023-q4": {
+                            "value": null,
+                            "period_type": "quarter"
+                        }
+                    },
+                    "activity_name": "Machinery Fuel",
+                    "activity_description": "Forklifts, excavators, cranes, other heavy machinery."
+                },
+                "cloud_computing": {
+                    "activity_name": "Cloud Computing",
+                    "activity_description": "Server & SaaS usage. "
+                },
+                "business_supplies": {
+                    "activity_name": "Business Supplies",
+                    "activity_description": "Consumables and single-use items: Inks, papers, napkins & towels, etc."
+                },
+                "business_equipment": {
+                    "activity_name": "Business Equipment",
+                    "activity_description": "Printers, sinks & toilets, computers, printers, other electronic appliances. "
+                },
+                "machinery_electricity": {
+                    "activity_name": "Machinery Electricity",
+                    "activity_description": "Forklifts, excavators, cranes, other heavy machinery. "
+                },
+                "professional_services": {
+                    "activity_name": "Professional Services",
+                    "activity_description": "Creative, marketing and communications, consultants, legal, etc. "
+                }
+            },
+            "journey_score": null,
+            "subcategory_name": "Operations"
+        }
+    },
+    "weighted_journey_score": null
+  };
+}